Remove locked revisions from the Cargo.toml
Locked revisions inside of a manifest generally indicate that a library is only
compatible with exactly one revision of the dependent library. In general these
locked revisions belong in a Cargo.lock instead of a Cargo.toml.

By including a locked revision, the dependency graph is forked such that
postgres will have its own copy of all of its dependencies, while all other
users of, for example, rust-openssl will get a different copy (the sources are
different).

This package is in general compatible with the current master branch of its
dependencies, and applications will have locked versions of all these
dependencies regardless.